requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
dateparser==1.1.8
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
psycopg2==2.9.6
pycryptodome==3.18.0
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
